ensembledatademo/pyproject.toml
[project]
name = "ensembledatademo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "ensembledata",
    "requests>=2.32.5",
    "pydantic-ai>=0.8.0",
    "openai>=1.102.0",
]
